+++ packages/ai/src/registry/provider-registry.test.ts
@@ -1,5 +1,5 @@
 import { NoSuchModelError } from '@ai-sdk/provider';
-import { MockEmbeddingModelV2 } from '../test/mock-embedding-model-v2';
+import { MockEmbeddingModelV3 } from '../test/mock-embedding-model-v2';
 import { MockLanguageModelV2 } from '../test/mock-language-model-v2';
 import { NoSuchProviderError } from './no-such-provider-error';
 import { createProviderRegistry } from './provider-registry';
@@ -159,7 +159,7 @@ describe('languageModel', () => {
 
 describe('textEmbeddingModel', () => {
   it('should return embedding model from provider using textEmbeddingModel', () => {
-    const model = new MockEmbeddingModelV2<string>();
+    const model = new MockEmbeddingModelV3<string>();
 
     const modelRegistry = createProviderRegistry({
       provider: {
@@ -224,7 +224,7 @@ describe('textEmbeddingModel', () => {
   });
 
   it('should support custom separator', () => {
-    const model = new MockEmbeddingModelV2<string>();
+    const model = new MockEmbeddingModelV3<string>();
 
     const modelRegistry = createProviderRegistry(
       {