+++ packages/ai/test/index.ts
@@ -4,7 +4,7 @@ export {
   convertReadableStreamToArray,
   mockId,
 } from '@ai-sdk/provider-utils/test';
-export { MockEmbeddingModelV2 } from '../src/test/mock-embedding-model-v2';
+export { MockEmbeddingModelV3 } from '../src/test/mock-embedding-model-v2';
 export { MockImageModelV3 } from '../src/test/mock-image-model-v2';
 export { MockLanguageModelV2 } from '../src/test/mock-language-model-v2';
 export { MockProviderV2 } from '../src/test/mock-provider-v2';