+++ packages/amazon-bedrock/src/bedrock-provider.ts
@@ -1,5 +1,5 @@
 import {
-  EmbeddingModelV2,
+  EmbeddingModelV3,
   ImageModelV3,
   LanguageModelV2,
   ProviderV2,
@@ -107,7 +107,7 @@ export interface AmazonBedrockProvider extends ProviderV2 {
 
   languageModel(modelId: BedrockChatModelId): LanguageModelV2;
 
-  embedding(modelId: BedrockEmbeddingModelId): EmbeddingModelV2<string>;
+  embedding(modelId: BedrockEmbeddingModelId): EmbeddingModelV3<string>;
 
   /**
 Creates a model for image generation.