+++ packages/provider/src/provider/v2/provider-v2.ts
@@ -1,4 +1,4 @@
-import { EmbeddingModelV2 } from '../../embedding-model/v2/embedding-model-v2';
+import { EmbeddingModelV3 } from '../../embedding-model/v2/embedding-model-v2';
 import { ImageModelV3 } from '../../image-model/v3/image-model-v3';
 import { LanguageModelV2 } from '../../language-model/v2/language-model-v2';
 import { SpeechModelV2 } from '../../speech-model/v2/speech-model-v2';
@@ -30,7 +30,7 @@ The model id is then passed to the provider function to get the model.
 
 @throws {NoSuchModelError} If no such model exists.
    */
-  textEmbeddingModel(modelId: string): EmbeddingModelV2<string>;
+  textEmbeddingModel(modelId: string): EmbeddingModelV3<string>;
 
   /**
 Returns the image model with the given id.