+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,18 @@
 
 ### Patch Changes
 
+- ai@5.1.0-beta.5
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@5.1.0-beta.4
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - ai@5.1.0-beta.3
 
 ## 0.0.1-beta.0